requirements.txt
# AP Job Posting Lead Generation System
# Core
python-dotenv==1.1.0
requests==2.32.3

# SerpAPI
google-search-results==2.4.2

# Google Sheets
gspread==6.1.4
google-auth==2.38.0

# Direct scraping (Tier 3 - niche boards)
playwright==1.50.0

# Data processing
beautifulsoup4==4.13.3

# Scheduling
schedule==1.2.2

# Dashboard
flask==3.1.0
gunicorn==23.0.0

# Cloud Database (Neon PostgreSQL)
psycopg2-binary==2.9.10

